requirements.txt
future
flask
gevent
requests
socketIO-client-2
pydvbcss
xmldiff
